ProcessMetrics, Registry Reset Undo - Added module process_metrics.py containing class ProcessMetrics, - ProcessMetrics class contains all process metrics related data, - added var process_metrics which is a ProcessMetrics object, which will collect process metrics if parameter collect_process_metrics is set to  true, - updated tests accordingly, - Removed 'Registry reset when process id changes' feature, to ensure that no points are missed, - removed tests for registry reset.
