+++ apptuit/pyformance/__init__.py
@@ -2,5 +2,7 @@
     A Pyformance Reporter for Apptuit
 """
 from .apptuit_reporter import ApptuitReporter
+from .process_metrics import ProcessMetrics
 
-__all__ = ['ApptuitReporter']
+__all__ = ['ApptuitReporter',
+           'ProcessMetrics']
