+++ apptuit/pyformance/apptuit_reporter.py
@@ -2,17 +2,15 @@
 """
 Apptuit Pyformance Reporter
 """
-import gc
 import os
-import resource
 import socket
 import sys
-import threading
 
 from pyformance import MetricsRegistry
 from pyformance.reporters.reporter import Reporter
 
 from apptuit.apptuit_client import Apptuit, DataPoint, ApptuitSendException, TimeSeriesName
+from .process_metrics import ProcessMetrics
 from ..utils import _get_tags_from_environment, strtobool
 
 NUMBER_OF_TOTAL_POINTS = "apptuit.reporter.send.total"
@@ -21,15 +19,6 @@ NUMBER_OF_FAILED_POINTS = "apptuit.reporter.send.failed"
 API_CALL_TIMER = "apptuit.reporter.send.time"
 DISABLE_HOST_TAG = "APPTUIT_DISABLE_HOST_TAG"
 BATCH_SIZE = 50000
-RESOURCE_STRUCT_RUSAGE = ["ru_utime", "ru_stime",
-                          "ru_maxrss", "ru_ixrss",
-                          "ru_idrss", "ru_isrss",
-                          "ru_minflt", "ru_majflt",
-                          "ru_nswap", "ru_inblock",
-                          "ru_oublock", "ru_msgsnd",
-                          "ru_msgrcv", "ru_nsignals",
-                          "ru_nvcsw", "ru_nivcsw"
-                          ]
 
 
 def default_error_handler(status_code, successful, failed, errors):
@@ -56,84 +45,6 @@ class ApptuitReporter(Reporter):
         primitives, such as meter, counter, gauge, etc., for collecting
         data and reports them asynchronously to Apptuit.
     """
-
-    def _get_gc_metric_names(self):
-        """
-        To get a list of gc metric names.
-        :return: a list of gc metric names.
-        """
-        gc_metric_names = [
-            TimeSeriesName.encode_metric("python.gc.collection",
-                                         {"type": "collection_0", "worker_id": self.pid}),
-            TimeSeriesName.encode_metric("python.gc.collection",
-                                         {"type": "collection_1", "worker_id": self.pid}),
-            TimeSeriesName.encode_metric("python.gc.collection",
-                                         {"type": "collection_2", "worker_id": self.pid}),
-            TimeSeriesName.encode_metric("python.gc.threshold",
-                                         {"type": "threshold_0", "worker_id": self.pid}),
-            TimeSeriesName.encode_metric("python.gc.threshold",
-                                         {"type": "threshold_1", "worker_id": self.pid}),
-            TimeSeriesName.encode_metric("python.gc.threshold",
-                                         {"type": "threshold_2", "worker_id": self.pid})
-        ]
-        return gc_metric_names
-
-    def _get_thread_metic_names(self):
-        """
-        To get a list of thread metric names.
-        :return: a list of thread metric names.
-        """
-        thread_metric_names = [
-            TimeSeriesName.encode_metric("python.thread",
-                                         {"type": "demon", "worker_id": self.pid}),
-            TimeSeriesName.encode_metric("python.thread",
-                                         {"type": "alive", "worker_id": self.pid}),
-            TimeSeriesName.encode_metric("python.thread",
-                                         {"type": "dummy", "worker_id": self.pid}),
-        ]
-        return thread_metric_names
-
-    def _get_resource_metic_names(self):
-        """
-        To get a list of resource metric names.
-        :return: a list of resource metric names.
-        """
-        resource_metric_names = [
-            TimeSeriesName.encode_metric("python.cpu.time.used.seconds",
-                                         {"type": "user", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.cpu.time.used.seconds",
-                                         {"type": "system", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.memory.usage.bytes",
-                                         {"type": "main", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.memory.usage.bytes",
-                                         {"type": "shared", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.memory.usage.bytes",
-                                         {"type": "unshared", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.memory.usage.bytes",
-                                         {"type": "unshared_stack_size",
-                                          "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.page.faults",
-                                         {"type": "major", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.page.faults",
-                                         {"type": "minor", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.process.swaps",
-                                         {"worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.block.operations",
-                                         {"type": "input", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.block.operations",
-                                         {"type": "output", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.ipc.messages",
-                                         {"type": "sent", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.ipc.messages",
-                                         {"type": "received", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.system.signals",
-                                         {"worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.context.switches",
-                                         {"type": "voluntary", "worker_id": self.pid, }),
-            TimeSeriesName.encode_metric("python.context.switches",
-                                         {"type": "involuntary", "worker_id": self.pid, }),
-        ]
-        return resource_metric_names
 
     def __init__(self, registry=None, reporting_interval=10, token=None,
                  api_endpoint="https://api.apptuit.ai", prefix="", tags=None,
@@ -194,13 +105,9 @@ class ApptuitReporter(Reporter):
         self._meta_metrics_registry = MetricsRegistry()
         self.error_handler = error_handler
         self.pid = os.getpid()
-        self.collect_process_metrics = collect_process_metrics
-        if self.collect_process_metrics:
-            self.resource_metric_names = self._get_resource_metic_names()
-            self.thread_metrics_names = self._get_thread_metic_names()
-            self.gc_metric_names = self._get_gc_metric_names()
-            self.previous_resource_metrics = [0] * len(self.resource_metric_names)
-            self.previous_gc_metrics = [0] * len(self.gc_metric_names)
+        self.process_metrics = None
+        if collect_process_metrics:
+            self.process_metrics = ProcessMetrics(self.pid, self.registry)
 
     def _update_counter(self, key, value):
         """
@@ -210,58 +117,6 @@ class ApptuitReporter(Reporter):
         """
         self._meta_metrics_registry.counter(key).inc(value)
 
-    def _collect_counter_from_list(self, metric_names, metric_values):
-        """
-        To increment list of counters `metric_names` with values `metric_values`.
-        :param metric_names: A list of counter names.
-        :param metric_values: A list of corresponding value to increment.
-        """
-        for ind, metric_value in enumerate(metric_values):
-            metric_counter = self.registry.counter(metric_names[ind])
-            metric_counter.inc(metric_value)
-
-    def _collect_gauge_from_list(self, metric_names, metric_values):
-        """
-        To increment list of gauge `metric_names` with values `metric_values`.
-        :param metric_names: A list of gauge names.
-        :param metric_values: A list of corresponding value to set.
-        """
-        for ind, metric in enumerate(metric_values):
-            metric_counter = self.registry.gauge(metric_names[ind])
-            metric_counter.set_value(metric)
-
-    def _collect_process_metrics(self):
-        """
-        To collect all the process metrics.
-        """
-        rm = resource.getrusage(resource.RUSAGE_SELF)
-        resource_metrics = {}
-        for res_name in RESOURCE_STRUCT_RUSAGE:
-            resource_metrics[res_name] = getattr(rm, res_name)
-        for res_name in RESOURCE_STRUCT_RUSAGE[2:6]:
-            resource_metrics[res_name] = resource_metrics[res_name] * 1024
-        resource_metrics = [cur_val - pre_val
-                            for cur_val, pre_val in
-                            zip([resource_metrics[key]
-                                 for key in resource_metrics], self.previous_resource_metrics)]
-        self._collect_counter_from_list(self.resource_metric_names, resource_metrics)
-        th_values = threading.enumerate()
-        thread_metrics = [
-            [t.daemon is True for t in th_values].count(True),
-            [t.daemon is False for t in th_values].count(True),
-            [isinstance(t, threading._DummyThread) for t in th_values].count(True)
-        ]
-        self._collect_gauge_from_list(self.thread_metrics_names, thread_metrics)
-        if gc.isenabled():
-            collection = list(gc.get_count())
-            threshold = list(gc.get_threshold())
-            gc_metrics = collection + threshold
-            gc_metrics = [cur_val - pre_val
-                          for cur_val, pre_val in zip(gc_metrics, self.previous_gc_metrics)]
-            self._collect_counter_from_list(self.gc_metric_names, gc_metrics)
-            self.previous_gc_metrics = gc_metrics
-        self.previous_resource_metrics = resource_metrics
-
     def report_now(self, registry=None, timestamp=None):
         """
         Report the data
@@ -269,19 +124,8 @@ class ApptuitReporter(Reporter):
             registry: pyformance Registry containing all metrics
             timestamp: timestamp of the data point
         """
-        pid = os.getpid()
-        if pid != self.pid:
-            self.registry.clear()
-            self.pid = pid
-            if self.collect_process_metrics:
-                self.resource_metric_names = self._get_resource_metic_names()
-                self.thread_metrics_names = self._get_thread_metic_names()
-                self.gc_metric_names = self._get_gc_metric_names()
-                self.previous_resource_metrics = [0] * len(self.resource_metric_names)
-                self.previous_gc_metrics = [0] * len(self.gc_metric_names)
-            return
-        if self.collect_process_metrics:
-            self._collect_process_metrics()
+        if self.process_metrics:
+            self.process_metrics.collect_process_metrics()
         dps = self._collect_data_points(registry or self.registry, timestamp)
         meta_dps = self._collect_data_points(self._meta_metrics_registry)
         if not dps:
